Refactor result generation:
- In clock.result:
  - rename ResultFactory hierarchy to Result,
  - remove its result method (which is framework-specific), and
  - move the hierarchy up one level to result package,
  - remove also ResultGenerator.
- Create new clock.bot.inline.query.result package with
  framework-specific result logic moved out of clock.result:
  - A InlineResult.from_result function that implements the
    previous result() function in ResultFormatter
  - A InlineResultGenerator class that resembles the previous
    ResultGenerator class and returns and generates the inline results
- Update InlineQueryAction to use new InlineResultGenerator
