+++ clock/bot/inline/query/action.py
@@ -2,11 +2,11 @@ from babel import Locale, UnknownLocaleError
 from bot.action.core.action import Action
 from bot.multithreading.work import Work
 
+from clock.bot.inline.query.result.generator import InlineResultGenerator
 from clock.bot.locale_cache import LocaleCache
 from clock.domain.time import TimePoint
 from clock.finder.api import ZoneFinderApi
 from clock.log.api import LogApi
-from clock.result.generator import ResultGenerator
 from clock.storage.api import StorageApi
 
 
@@ -41,7 +41,7 @@ class InlineQueryClockAction(Action):
         offset_end = offset + MAX_RESULTS_PER_QUERY
         next_offset = self.__get_next_offset(len(zones), offset_end)
 
-        results = ResultGenerator.generate(current_time, locale, zones[offset:offset_end])
+        results = InlineResultGenerator.generate(current_time, locale, zones[offset:offset_end])
 
         processing_time = TimePoint.current_timestamp() - current_time.timestamp
