+++ clock/result/__init__.py
@@ -1,0 +1,15 @@
+class Result:
+    def id(self):
+        raise NotImplementedError()
+
+    def title(self):
+        raise NotImplementedError()
+
+    def description(self):
+        raise NotImplementedError()
+
+    def message(self):
+        raise NotImplementedError()
+
+    def image_url(self):
+        raise NotImplementedError()
