+++ clock/result/factory.py
@@ -6,11 +6,11 @@ from clock.domain.country import Country, CountryFormatter
 from clock.domain.datetimezone import DateTimeZone, DateTimeZoneFormatter
 from clock.domain.time import TimePoint
 from clock.domain.zone import Zone
-from clock.result.formatter.country import CountryResultFormatter
-from clock.result.formatter.date_time_zone import DateTimeZoneResultFormatter
+from clock.result.country import CountryResult
+from clock.result.date_time_zone import DateTimeZoneResult
 
 
-class ResultFormatterFactory:
+class ResultFactory:
     @classmethod
     def get(cls, time_point: TimePoint, zone: Union[Zone, Country], locale: Locale):
         if isinstance(zone, Country):
@@ -21,7 +21,7 @@ class ResultFormatterFactory:
     @classmethod
     def get_for_zone(cls, time_point: TimePoint, zone: Zone, locale: Locale):
         date_time_zone_formatter = cls._get_date_time_zone_formatter(time_point, zone, locale)
-        return DateTimeZoneResultFormatter(date_time_zone_formatter)
+        return DateTimeZoneResult(date_time_zone_formatter)
 
     @staticmethod
     def _get_date_time_zone_formatter(time_point: TimePoint, zone: Zone, locale: Locale):
@@ -33,4 +33,4 @@ class ResultFormatterFactory:
         country_formatter = CountryFormatter(country, locale)
         date_time_zone_formatters = [cls._get_date_time_zone_formatter(time_point, zone, locale)
                                      for zone in country.zones]
-        return CountryResultFormatter(time_point, country_formatter, date_time_zone_formatters)
+        return CountryResult(time_point, country_formatter, date_time_zone_formatters)
